Bugfix: wrong cast
git-svn-id: file:///home/code/svn/xrdp/trunk@781 c7ce0e6c-3862-4c07-88d1-63b9e066504f

+++ common/parse.h
@@ -314,7 +314,7 @@ struct stream
 #else
 #define out_uint64_le(s, v) \
 { \
-  *((unsigned int*)((s)->p)) = (v); \
+  *((uint64_t*)((s)->p)) = (v); \
   (s)->p += 8; \
 }
 #endif
